[KMK003] update infra FLEX MODE

+++ nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.infra/src/main/java/nts/uk/ctx/at/shared/infra/entity/worktime/KshmtFlexOtTimeSet.java
@@ -5,6 +5,7 @@
 package nts.uk.ctx.at.shared.infra.entity.worktime;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 import javax.persistence.Column;
 import javax.persistence.EmbeddedId;
@@ -61,11 +62,11 @@ public class KshmtFlexOtTimeSet extends UkJpaEntity implements Serializable {
 
 	/** The ot frame no. */
 	@Column(name = "OT_FRAME_NO")
-	private int otFrameNo;
+	private BigDecimal otFrameNo;
 
 	/** The legal ot frame no. */
 	@Column(name = "LEGAL_OT_FRAME_NO")
-	private Integer legalOtFrameNo;
+	private BigDecimal legalOtFrameNo;
 
 	/** The payoff order. */
 	@Column(name = "PAYOFF_ORDER")
